requirements.dev.txt
-r requirements.txt

django-browser-reload==1.12.1
